Merge default-rocks plugin into this one
I would have preferred for the default-rocks plugin to be separate, but
there doesn't appear to be a reliable way to ensure
`install_default_rocks` will run after `install_luarocks`. The only way
I could think of to handle it is to specify in the README that the name
used in the plugin folder must come after this one, sorted naturally.
This seems like it would cause a lot of confusion, so I'm just combining
them, and making the default-rocks functionality optional.